Split to smaller files
