+++ lib/fwupd.dart
@@ -1,1 +1,6 @@
 export 'src/fwupd_client.dart';
+export 'src/fwupd_device.dart';
+export 'src/fwupd_exceptions.dart';
+export 'src/fwupd_plugin.dart';
+export 'src/fwupd_release.dart';
+export 'src/fwupd_remote.dart';
