+++ lib/src/fwupd_client.dart
@@ -3,7 +3,12 @@ import 'dart:io';
 
 import 'package:collection/collection.dart';
 import 'package:dbus/dbus.dart';
-import 'package:meta/meta.dart';
+
+import 'fwupd_exceptions.dart';
+import 'fwupd_device.dart';
+import 'fwupd_plugin.dart';
+import 'fwupd_release.dart';
+import 'fwupd_remote.dart';
 
 enum FwupdStatus {
   unknown,
@@ -22,92 +27,6 @@ enum FwupdStatus {
   shutdown
 }
 
-enum FwupdDeviceFlag {
-  internal,
-  updatable,
-  onlyOffline,
-  requireAc,
-  locked,
-  supported,
-  needsBootloader,
-  registered,
-  needsReboot,
-  reported,
-  notified,
-  useRuntimeVersion,
-  installParentFirst,
-  isBootloader,
-  waitForReplug,
-  ignoreValidation,
-  trusted,
-  needsShutdown,
-  anotherWriteRequired,
-  noAutoInstanceIds,
-  needsActivation,
-  ensureSemver,
-  historical,
-  onlySupported,
-  willDisappear,
-  canVerify,
-  canVerifyImage,
-  dualImage,
-  selfRecovery,
-  usableDuringUpdate,
-  versionCheckRequired,
-  installAllReleases,
-  mdSetName,
-  mdSetNameCategory,
-  mdSetVerfmt,
-  addCounterpartGuids,
-  noGuidMatching,
-  updatableHidden,
-  skipsRestart,
-  hasMultipleBranches,
-  backupBeforeInstall,
-  mdSetIcon,
-  wildcardInstall,
-  onlyVersionUpgrade,
-  unreachable,
-  affectsFde
-}
-
-enum FwupdVersionFormat {
-  unknown,
-  plain,
-  number,
-  pair,
-  triplet,
-  quad,
-  bcd,
-  intelMe,
-  intelMe2,
-  surfaceLegacy,
-  surface,
-  dellBios,
-  hex
-}
-
-enum FwupdReleaseFlag {
-  trustedPayload,
-  trustedMetadata,
-  isUpgrade,
-  isDowngrade,
-  blockedVersion,
-  blockedApproval,
-  isAlternateBranch
-}
-
-enum FwupdReleaseUrgency { unknown, low, medium, high, critical }
-
-enum FwupdUpdateState {
-  unknown,
-  pending,
-  success,
-  failed,
-  reboot,
-  failedTransient,
-}
-
 enum FwupdInstallFlag {
   offline,
   allowReinstall,
@@ -116,362 +35,6 @@ enum FwupdInstallFlag {
   noHistory,
   allowBranchSwitch,
   ignorePower,
-}
-
-class FwupdException implements Exception {
-  final dynamic message;
-
-  const FwupdException([this.message]);
-
-  @override
-  String toString() => '$runtimeType: $message';
-}
-
-/// Internal error
-class FwupdInternalException extends FwupdException {
-  const FwupdInternalException([dynamic message]) : super(message);
-}
-
-/// Installed newer firmware version
-class FwupdVersionNewerException extends FwupdException {
-  const FwupdVersionNewerException([dynamic message]) : super(message);
-}
-
-/// Installed same firmware version
-class FwupdVersionSameException extends FwupdException {
-  const FwupdVersionSameException([dynamic message]) : super(message);
-}
-
-/// Already set be be installed offline
-class FwupdAlreadyPendingException extends FwupdException {
-  const FwupdAlreadyPendingException([dynamic message]) : super(message);
-}
-
-/// Failed to get authentication
-class FwupdAuthFailedException extends FwupdException {
-  const FwupdAuthFailedException([dynamic message]) : super(message);
-}
-
-/// Failed to read from device
-class FwupdReadException extends FwupdException {
-  const FwupdReadException([dynamic message]) : super(message);
-}
-
-/// Failed to write to the device
-class FwupdWriteException extends FwupdException {
-  const FwupdWriteException([dynamic message]) : super(message);
-}
-
-/// Invalid file format
-class FwupdInvalidFileException extends FwupdException {
-  const FwupdInvalidFileException([dynamic message]) : super(message);
-}
-
-/// No matching device exists
-class FwupdNotFoundException extends FwupdException {
-  const FwupdNotFoundException([dynamic message]) : super(message);
-}
-
-/// Nothing to do
-class FwupdNothingToDoException extends FwupdException {
-  const FwupdNothingToDoException([dynamic message]) : super(message);
-}
-
-/// Action was not possible
-class FwupdNotSupportedException extends FwupdException {
-  const FwupdNotSupportedException([dynamic message]) : super(message);
-}
-
-/// Signature was invalid
-class FwupdSignatureInvalidException extends FwupdException {
-  const FwupdSignatureInvalidException([dynamic message]) : super(message);
-}
-
-/// AC power was required
-class FwupdAcPowerRequiredException extends FwupdException {
-  const FwupdAcPowerRequiredException([dynamic message]) : super(message);
-}
-
-/// Permission was denied
-class FwupdPermissionDeniedException extends FwupdException {
-  const FwupdPermissionDeniedException([dynamic message]) : super(message);
-}
-
-/// User has configured their system in a broken way
-class FwupdBrokenSystemException extends FwupdException {
-  const FwupdBrokenSystemException([dynamic message]) : super(message);
-}
-
-/// The system battery level is too low
-class FwupdBatteryLevelTooLowException extends FwupdException {
-  const FwupdBatteryLevelTooLowException([dynamic message]) : super(message);
-}
-
-/// User needs to do an action to complete the update
-class FwupdNeedsUserActionException extends FwupdException {
-  const FwupdNeedsUserActionException([dynamic message]) : super(message);
-}
-
-@immutable
-class FwupdDevice {
-  final String? checksum;
-  final DateTime? created;
-  final String deviceId;
-  final Set<FwupdDeviceFlag> flags;
-  final List<String> guid;
-  final List<String> icon;
-  final DateTime? modified;
-  final String name;
-  final String? parentDeviceId;
-  final String plugin;
-  final String? protocol;
-  final String? summary;
-  final FwupdUpdateState updateState;
-  final String? vendor;
-  final String? vendorId;
-  final String? version;
-  final String? versionBootloader;
-  final FwupdVersionFormat versionFormat;
-  final String? versionLowest;
-
-  FwupdDevice(
-      {this.checksum,
-      this.created,
-      required this.deviceId,
-      this.flags = const {},
-      this.guid = const [],
-      this.icon = const [],
-      this.modified,
-      required this.name,
-      this.parentDeviceId,
-      required this.plugin,
-      this.protocol,
-      this.summary,
-      this.updateState = FwupdUpdateState.unknown,
-      this.vendor,
-      this.vendorId,
-      this.version,
-      this.versionBootloader,
-      this.versionFormat = FwupdVersionFormat.unknown,
-      this.versionLowest});
-
-  @override
-  String toString() =>
-      "FwupdDevice(checksum: $checksum, created: $created, deviceId: $deviceId, flags: $flags, guid: $guid, icon: $icon, modified: $modified, name: '$name', parentDeviceId: $parentDeviceId, plugin: $plugin, protocol: $protocol, summary: $summary, updateState: $updateState, vendor: $vendor, vendorId: $vendorId, version: $version, versionBootloader: $versionBootloader, versionFormat: $versionFormat, versionLowest: $versionLowest)";
-
-  @override
-  int get hashCode => Object.hashAll([
-        checksum,
-        created,
-        deviceId,
-        Object.hashAll(flags),
-        Object.hashAll(guid),
-        Object.hashAll(icon),
-        modified,
-        name,
-        parentDeviceId,
-        plugin,
-        protocol,
-        summary,
-        updateState,
-        vendor,
-        vendorId,
-        version,
-        versionBootloader,
-        versionFormat,
-        versionLowest
-      ]);
-
-  @override
-  bool operator ==(Object other) {
-    if (identical(this, other)) return true;
-    if (other.runtimeType != runtimeType) return false;
-    return other is FwupdDevice &&
-        other.checksum == checksum &&
-        other.created == created &&
-        other.deviceId == deviceId &&
-        const SetEquality<FwupdDeviceFlag>().equals(other.flags, flags) &&
-        const ListEquality<String>().equals(other.guid, guid) &&
-        const ListEquality<String>().equals(other.icon, icon) &&
-        other.modified == modified &&
-        other.name == name &&
-        other.parentDeviceId == parentDeviceId &&
-        other.plugin == plugin &&
-        other.protocol == protocol &&
-        other.summary == summary &&
-        other.updateState == updateState &&
-        other.vendor == vendor &&
-        other.vendorId == vendorId &&
-        other.version == version &&
-        other.versionBootloader == versionBootloader &&
-        other.versionFormat == versionFormat &&
-        other.versionLowest == versionLowest;
-  }
-}
-
-class FwupdPlugin {
-  final String name;
-
-  FwupdPlugin({
-    required this.name,
-  });
-
-  @override
-  String toString() => 'FwupdDevice(name: $name)';
-}
-
-@immutable
-class FwupdRelease {
-  final String? appstreamId;
-  final String? checksum;
-  final DateTime? created;
-  final String description;
-  final String? filename;
-  final String homepage;
-  final int installDuration;
-  final String license;
-  final List<String> locations;
-  final String name;
-  final String? protocol;
-  final String? remoteId;
-  final int size;
-  final String summary;
-  final Set<FwupdReleaseFlag> flags;
-  final FwupdReleaseUrgency urgency;
-  final String? uri;
-  final String vendor;
-  final String version;
-
-  FwupdRelease(
-      {this.appstreamId,
-      this.checksum,
-      this.created,
-      this.description = '',
-      this.filename,
-      this.homepage = '',
-      this.installDuration = 0,
-      this.license = '',
-      this.locations = const [],
-      required this.name,
-      this.protocol,
-      this.remoteId,
-      this.size = 0,
-      this.summary = '',
-      this.flags = const {},
-      this.urgency = FwupdReleaseUrgency.unknown,
-      this.uri,
-      this.vendor = '',
-      this.version = ''});
-
-  @override
-  String toString() =>
-      "FwupdRelease(appstreamId: $appstreamId, checksum: $checksum, created: $created, description: '$description', filename: $filename, homepage: $homepage, license: $license, locations: $locations, name: '$name', protocol: $protocol, remoteId: $remoteId, size: $size, flags: $flags, urgency: $urgency, uri: $uri, vendor: '$vendor', version: '$version')";
-
-  @override
-  int get hashCode => Object.hashAll([
-        appstreamId,
-        checksum,
-        created,
-        description,
-        filename,
-        homepage,
-        installDuration,
-        license,
-        Object.hashAll(locations),
-        name,
-        protocol,
-        remoteId,
-        size,
-        summary,
-        Object.hashAll(flags),
-        urgency,
-        uri,
-        vendor,
-        version
-      ]);
-
-  @override
-  bool operator ==(Object other) {
-    if (identical(this, other)) return true;
-    if (other.runtimeType != runtimeType) return false;
-    return other is FwupdRelease &&
-        runtimeType == other.runtimeType &&
-        appstreamId == other.appstreamId &&
-        checksum == other.checksum &&
-        created == other.created &&
-        description == other.description &&
-        filename == other.filename &&
-        homepage == other.homepage &&
-        installDuration == other.installDuration &&
-        license == other.license &&
-        const ListEquality<String>().equals(locations, other.locations) &&
-        name == other.name &&
-        protocol == other.protocol &&
-        remoteId == other.remoteId &&
-        size == other.size &&
-        summary == other.summary &&
-        const SetEquality<FwupdReleaseFlag>().equals(flags, other.flags) &&
-        urgency == other.urgency &&
-        uri == other.uri &&
-        vendor == other.vendor &&
-        version == other.version;
-  }
-}
-
-enum FwupdRemoteKind { unknown, download, local, directory }
-enum FwupdKeyringKind { unknown, none, gpg, pkcs7, jcat }
-
-class FwupdRemote {
-  final DateTime? age;
-  final String? agreement;
-  final bool approvalRequired;
-  final bool automaticReports;
-  final bool automaticSecurityReports;
-  final String? checksum;
-  final bool enabled;
-  final String? filenameCache;
-  final String? filenameCacheSig;
-  final String? filenameSource;
-  final String? firmwareBaseUri;
-  final String id;
-  final FwupdKeyringKind keyringKind;
-  final FwupdRemoteKind kind;
-  final String? metadataUri;
-  final String? password;
-  final int priority;
-  final String? remotesDir;
-  final String? reportUri;
-  final String? securityReportUri;
-  final String? title;
-  final String? username;
-
-  FwupdRemote(
-      {this.age,
-      this.agreement,
-      this.approvalRequired = false,
-      this.automaticReports = false,
-      this.automaticSecurityReports = false,
-      this.checksum,
-      this.enabled = false,
-      this.filenameCache,
-      this.filenameCacheSig,
-      this.filenameSource,
-      this.firmwareBaseUri,
-      required this.id,
-      this.keyringKind = FwupdKeyringKind.jcat,
-      this.kind = FwupdRemoteKind.unknown,
-      this.metadataUri,
-      this.password,
-      this.priority = 0,
-      this.remotesDir,
-      this.reportUri,
-      this.securityReportUri,
-      this.title,
-      this.username});
-
-  @override
-  String toString() =>
-      "FwupdRemote(age: $age, agreement: '$agreement', approvalRequired: $approvalRequired, automaticReports: $automaticReports, automaticSecurityReports: $automaticSecurityReports, checksum: '$checksum', enabled: $enabled, filenameCache: '$filenameCache', filenameCacheSig: '$filenameCacheSig', filenameSource: '$filenameSource', firmwareBaseUri: '$firmwareBaseUri', id: '$id', keyringKind: $keyringKind, kind: $kind, metadataUri: '$metadataUri', password: '${password?.replaceAll(RegExp('.'), '*')}', priority: $priority, remotesDir: '$remotesDir', reportUri: '$reportUri', securityReportUri: '$securityReportUri', title: '$title', username: '$username')";
 }
 
 /// A client that connects to fwupd.
