+++ lib/src/fwupd_plugin.dart
@@ -1,0 +1,10 @@
+class FwupdPlugin {
+  final String name;
+
+  FwupdPlugin({
+    required this.name,
+  });
+
+  @override
+  String toString() => 'FwupdDevice(name: $name)';
+}
